requirements.txt
# PyTorch CPU version
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
opencv-python>=4.5.0
pandas>=1.3.0
openpyxl>=3.0.0
sqlalchemy>=1.4.0
pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.60.0
PyQt5>=5.15.0
